+++ packages/quantic/playwright/page-object/querySuggestObject.ts
@@ -1,4 +1,5 @@
 import {Page, Request, Response} from '@playwright/test';
+import {querySuggestResponses} from '@coveo/platform-mock-api/search';
 
 export class QuerySuggestObject {
   constructor(
@@ -36,10 +37,8 @@ export class QuerySuggestObject {
 
   async mockQuerySuggestResponse(suggestions: string[]) {
     await this.page.route(this.querySuggestRequestRegex, async (route) => {
-      const apiResponse = await this.page.request.fetch(route.request());
-      const originalBody = await apiResponse.json();
       const modifiedBody = {
-        ...originalBody,
+        ...querySuggestResponses.baseResponse,
         completions: suggestions.map((suggestion) => ({
           expression: suggestion,
           highlighted: `[${suggestion}]`,